added reserve option to client accepting and denying i no room

+++ delete_migrations.py
@@ -9,7 +9,7 @@ def delete_files_in_migrations(directory):
             migrations_dir = os.path.join(root, 'migrations')
             for filename in os.listdir(migrations_dir):
                 file_path = os.path.join(migrations_dir, filename)
-                if os.path.isfile(file_path) and filename != '__init__.py':
+                if os.path.isfile(file_path) and filename != '__init__.py' and filename != '0001_initial_data.py':
                     os.remove(file_path)
                     print(f"S'ha borrat {filename} del directori {migrations_dir}")
